requirement.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
pandas>=2.0.0
matplotlib>=3.8.0
aiosqlite>=0.19.0
